Remove baseline PNGs
